remove Enum override
