+++ src/python/pants/util/meta.py
@@ -3,8 +3,9 @@
 
 from abc import ABC, abstractmethod
 from dataclasses import FrozenInstanceError
+from enum import Enum
 from functools import wraps
-from typing import Any, Callable, Optional, Type, TypeVar, Union
+from typing import Any, Callable, Mapping, Optional, Type, TypeVar, Union, cast
 
 
 T = TypeVar("T")
